* fixed a compile-bug with FrBClass::typeID() that occured when compiling on a 64 bits computer
git-svn-id: file:///Users/ng/svn-local/to_share/frb/trunk@146 f04dad46-9b18-0410-952a-bb25a776dc73

+++ src/parser/frbclass.h
@@ -228,7 +228,7 @@ public:
       *                             // classA & classB are the same class
       *                         }
       */
-    inline unsigned int typeID() const { return reinterpret_cast<unsigned int>(this); }
+    inline long unsigned int typeID() const { return reinterpret_cast<unsigned long int>(this); }
    
     /** Return the full name of this class, ie Module1.Module2.MyClass1 */
     String fullName() const;
